requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dataclasses-json==0.5.9
exceptiongroup==1.1.2
fastapi==0.99.1
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
langchain==0.0.223
langchainplus-sdk==0.0.20
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openapi-schema-pydantic==1.2.4
packaging==23.1
pdfminer.six==20221105
portalocker==2.7.0
protobuf==4.23.3
pycparser==2.21
pydantic==1.10.11
PyYAML==6.0
qdrant-client==1.3.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tenacity==8.2.2
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==1.26.16
yarl==1.9.2
